requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
librosa==0.10.1
lightgbm==4.3.0
matplotlib==3.5.1
mysql_connector_repackaged==0.3.1
numpy==1.24.0
pandas==1.2.4
Pillow==9.4.0
Pillow==10.3.0
torch==2.0.1+cu117
torchvision==0.15.2+cu117
